docs(steering): add comprehensive API, Java, database, and core standards documentation
- Add API conventions guide covering URL design, validation, error codes, request records, boolean fields, and controller patterns
- Add core rules document establishing project-wide principles for code quality and consistency
- Add Java standards guide with formatting, naming, exception handling, and testing conventions
- Add database standards covering entity design, repository patterns, and query optimization
- Update all controllers (Address, Auth, Cart, Inventory, Order, Promotion, Review, Upload, Wishlist) to align with documented API conventions
- Add OrderDataSeeder and ReviewDataSeeder for comprehensive test data generation
- Enhance ReviewEntity with additional fields and validation
- Improve ProductRepositoryImpl query performance and filtering capabilities
- Update ProductSpecification with enhanced filtering logic
- Update ReviewJpaRepository and ReviewRepositoryImpl to support new seeding requirements
- Establish living documentation for team alignment and code consistency

+++ src/main/java/com/nitrotech/api/infrastructure/persistence/repository/ReviewRepositoryImpl.java
@@ -89,7 +89,7 @@ public class ReviewRepositoryImpl implements ReviewRepository {
         String userName = userJpa.findById(e.getUserId())
                 .map(UserEntity::getName).orElse(null);
         return new ReviewData(e.getId(), e.getProductId(), e.getUserId(), userName,
-                e.getOrderId(), e.getRating(), e.getComment(), e.getImages(),
+                e.getOrderId(), e.getRating().intValue(), e.getComment(), e.getImages(),
                 e.getStatus(), e.getCreatedAt(), e.getUpdatedAt());
     }
 }
